fix: controlling context menu in the `DocumentView`, refactored useEffect to a shared file
